add room and view all rooms successfully integrated
