+++ backend/src/v1/hotels/hotels.service.ts
@@ -7,7 +7,7 @@ import { NotFoundException } from '@nestjs/common';
 import { Hotel } from 'src/common/entities/hotel.entity';
 import { Food } from 'src/common/entities/food.entity';
 import { Booking } from 'src/common/entities/booking.entity';
-import { CreateOrderDto } from './dto/create-order.dto';
+import { CreateOrderDto } from '../food-menu/dto/create-order.dto';
 import { Order } from 'src/common/entities/order.entity';
 import { OrderItem } from 'src/common/entities/order-item.entity';
 
@@ -55,74 +55,6 @@ export class HotelService {
       }
       return hotels;  //Returns all hotels 
   }
-
-
-  async getAllFood(): Promise<Food[]> {
-    const foods = await this.foodRepository.find();  
-    if (foods.length == 0) {
-        throw new NotFoundException(`Foods not found`);  // Throws 'Foods not Found exception if there no hotel inside the database 
-      }
-      return foods ;  //Returns all Foods
-  }
-
-  async createOrder(createOrderDto: CreateOrderDto): Promise<{ orderId: string; totalPrice: number; message: string }> {
-    const { bookingId, items } = createOrderDto;
-
-    // Check if booking exists
-    const booking = await this.bookingRepository.findOne({ where: { id: bookingId } });
-    if (!booking) {
-      throw new NotFoundException('Booking not found.');
-    }
-
-    // Process food items
-    let totalPrice = 0;
-    const orderItems: OrderItem[] = [];
-
-    for (const item of items) {
-      const food = await this.foodRepository.findOne({ where: { id: item.foodId, status: 'Available' } });
-
-      if (!food) {
-        throw new NotFoundException(`Food item with ID ${item.foodId} not found or unavailable.`);
-      }
-
-      const itemTotal = food.price * item.quantity;
-      totalPrice += itemTotal;
-
-      const orderItem = this.orderItemRepository.create({
-        food,
-        quantity: item.quantity,
-        price: itemTotal,
-      });
-
-      orderItems.push(orderItem);
-    }
-
-    // Create order
-    const order = this.orderRepository.create({
-      booking,
-      items: orderItems,
-      totalPrice,
-      status: 'pending',
-    });
-
-    const savedOrder = await this.orderRepository.save(order);
-
-    // Save order items
-    for (const orderItem of orderItems) {
-      orderItem.order = savedOrder;
-      await this.orderItemRepository.save(orderItem);
-    }
-
-    return {
-      orderId: savedOrder.id,
-      totalPrice,
-      message: 'Food order placed successfully',
-    };
-  }
-
-
-
-
   
 }
